se agrega en el factory de ventas el cliente

+++ database/migrations/2019_02_14_171104_create_sales_details_table.php
@@ -16,7 +16,7 @@ class CreateSalesDetailsTable extends Migration
         Schema::create('sales_details', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('quantity');
-            $table->integer('amount');
+            $table->decimal('amount',18,2);
             $table->timestamps();
 
             $table->unsignedInteger('sale_id');
